+++ ui/game_updater_worker.py
